wip: updating integration tests to use comet rpc where required
