+++ core/fetcher_no_race_test.go
@@ -4,6 +4,7 @@ package core
 
 import (
 	"context"
+	"github.com/celestiaorg/celestia-node/internal/comet"
 	"testing"
 	"time"
 
@@ -20,7 +21,7 @@ func TestBlockFetcherHeaderValues(t *testing.T) {
 
 	cfg := DefaultTestConfig()
 	StartTestNodeWithConfig(t, cfg)
-	client, err := newCometGRPCConn(cfg.TmConfig.RPC.GRPCListenAddress)
+	client, err := comet.NewCometGRPCConn(cfg.TmConfig.RPC.GRPCListenAddress)
 	require.NoError(t, err)
 	fetcher, err := NewBlockFetcher(client)
 	require.NoError(t, err)
