+++ core/header_test.go
@@ -3,6 +3,7 @@ package core
 import (
 	"context"
 	"fmt"
+	"github.com/celestiaorg/celestia-node/internal/comet"
 	"testing"
 
 	"github.com/cometbft/cometbft/libs/rand"
@@ -23,8 +24,8 @@ func TestMakeExtendedHeaderForEmptyBlock(t *testing.T) {
 	cfg := DefaultTestConfig()
 	StartTestNodeWithConfig(t, cfg)
 
-	client, err := newCometGRPCConn(cfg.TmConfig.RPC.GRPCListenAddress)
-	
+	client, err := comet.NewCometGRPCConn(cfg.TmConfig.RPC.GRPCListenAddress)
+
 	require.NoError(t, err)
 	fetcher, err := NewBlockFetcher(client)
 	require.NoError(t, err)
